Adjust RENDER-RSS to handle a feed with no entry.
In case we have no actual content, to add to the feed, create an RSS feed without an <entry>.

Before, we were creating an RSS feed with an <entry> with empty content.

Set the date of the entry to be exactly midnight if we have no content, and midnight one second, if we have an entry;

This allows us to avoid duplicate posts of the RSS feed to slack for a given day, while still handling the case where
our first fetches of the day have no data, followed by a fetch with actual data.